Ekstraher oppsett for unleash i egen builder.
-Unngå exceptions ved kjøring av test/lokalt
-Eksplisitt sett FakeUnleash som default for testSusebakover()
